+++ common/src/main/java/com/lowdragmc/lowdraglib/gui/widget/SceneWidget.java
@@ -423,8 +423,8 @@ public class SceneWidget extends WidgetGroup {
     }
 
     @Override
-    public Object getIngredientOverMouse(double mouseX, double mouseY) {
-        Object result = super.getIngredientOverMouse(mouseX, mouseY);
+    public Object getXEIIngredientOverMouse(double mouseX, double mouseY) {
+        Object result = super.getXEIIngredientOverMouse(mouseX, mouseY);
         if (result == null && hoverItem != null && !hoverItem.isEmpty()) {
             if (LDLib.isJeiLoaded()) {
                 return JEIPlugin.getItemIngredient(hoverItem, (int) mouseX, (int) mouseY, 1, 1);