+++ common/src/main/java/com/lowdragmc/lowdraglib/jei/ModularUIRecipeCategory.java
@@ -5,14 +5,20 @@ import com.lowdragmc.lowdraglib.gui.widget.Widget;
 import lombok.val;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.builder.IRecipeSlotBuilder;
+import mezz.jei.api.ingredients.IIngredientType;
+import mezz.jei.api.ingredients.ITypedIngredient;
 import mezz.jei.api.recipe.IFocusGroup;
 import mezz.jei.api.recipe.RecipeIngredientRole;
 import mezz.jei.api.recipe.category.IRecipeCategory;
 import mezz.jei.api.runtime.IClickableIngredient;
 import net.minecraft.MethodsReturnNonnullByDefault;
 
+import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @author KilaBash
@@ -29,23 +35,36 @@ public abstract class ModularUIRecipeCategory<T extends ModularWrapper<?>> imple
         for (int i = 0; i < flatVisibleWidgetCollection.size(); i++) {
             Widget widget = flatVisibleWidgetCollection.get(i);
             if (widget instanceof IRecipeIngredientSlot slot) {
-                IRecipeSlotBuilder slotBuilder = builder.addSlot(mapToRole(slot.getIngredientIO()), i, -1);
-                Object ingredient = slot.getJEIIngredient();
-                if (ingredient instanceof IClickableIngredient<?> clickableIngredient) {
-                    val typedIngredient = clickableIngredient.getTypedIngredient();
-                    slotBuilder.addIngredientsUnsafe(List.of(typedIngredient.getIngredient()));
+                var role = mapToRole(slot.getIngredientIO());
+                if (role == null) { // both
+                    addJEISlot(builder, slot, RecipeIngredientRole.INPUT, i);
+                    addJEISlot(builder, slot, RecipeIngredientRole.OUTPUT, i);
+                } else {
+                    addJEISlot(builder, slot, role, i);
                 }
             }
         }
 
     }
 
+    private static void addJEISlot(IRecipeLayoutBuilder builder, IRecipeIngredientSlot slot, RecipeIngredientRole role, int index) {
+        IRecipeSlotBuilder slotBuilder = builder.addSlot(role, index, -1);
+        Map<IIngredientType, List> map = new HashMap<>();
+        slot.getXEIIngredients().stream()
+                .filter(IClickableIngredient.class::isInstance)
+                .map(IClickableIngredient.class::cast)
+                .forEach(clickableIngredient -> map.computeIfAbsent(clickableIngredient.getTypedIngredient().getType(), a -> new ArrayList<>()).add(clickableIngredient.getTypedIngredient().getIngredient()));
+        map.forEach((ingredient, list) -> slotBuilder.addIngredients(ingredient, list));
+    }
+
+    @Nullable
     private RecipeIngredientRole mapToRole(IngredientIO ingredientIO) {
         return switch (ingredientIO) {
             case INPUT -> RecipeIngredientRole.INPUT;
             case OUTPUT -> RecipeIngredientRole.OUTPUT;
             case CATALYST -> RecipeIngredientRole.CATALYST;
             case RENDER_ONLY -> RecipeIngredientRole.RENDER_ONLY;
+            case BOTH -> null;
         };
     }